fix: change CommonResult.code from long to int

+++ common/src/main/java/szu/common/api/CommonResult.java
@@ -8,7 +8,7 @@ public class CommonResult<T> {
     /**
      * 状态码
      */
-    private long code;
+    private int code;
     /**
      * 提示信息
      */
@@ -21,7 +21,7 @@ public class CommonResult<T> {
     protected CommonResult() {
     }
 
-    protected CommonResult(long code, String message, T data) {
+    protected CommonResult(int code, String message, T data) {
         this.code = code;
         this.message = message;
         this.data = data;
@@ -33,7 +33,7 @@ public class CommonResult<T> {
      * @param data 获取的数据
      */
     public static <T> CommonResult<T> success(T data) {
-        return new CommonResult<T>(ResultCode.SUCCESS.getCode(), ResultCode.SUCCESS.getMessage(), data);
+        return new CommonResult<T>(Long.valueOf(ResultCode.SUCCESS.getCode()).intValue(), ResultCode.SUCCESS.getMessage(), data);
     }
 
     /**
@@ -43,7 +43,7 @@ public class CommonResult<T> {
      * @param  message 提示信息
      */
     public static <T> CommonResult<T> success(T data, String message) {
-        return new CommonResult<T>(ResultCode.SUCCESS.getCode(), message, data);
+        return new CommonResult<T>(Long.valueOf(ResultCode.SUCCESS.getCode()).intValue(), message, data);
     }
 
     /**
@@ -51,7 +51,7 @@ public class CommonResult<T> {
      * @param errorCode 错误码
      */
     public static <T> CommonResult<T> failed(IErrorCode errorCode) {
-        return new CommonResult<T>(errorCode.getCode(), errorCode.getMessage(), null);
+        return new CommonResult<T>(Long.valueOf(errorCode.getCode()).intValue(), errorCode.getMessage(), null);
     }
 
     /**
@@ -60,7 +60,7 @@ public class CommonResult<T> {
      * @param message 错误信息
      */
     public static <T> CommonResult<T> failed(IErrorCode errorCode,String message) {
-        return new CommonResult<T>(errorCode.getCode(), message, null);
+        return new CommonResult<T>(Long.valueOf(errorCode.getCode()).intValue(), message, null);
     }
 
     /**
@@ -68,7 +68,7 @@ public class CommonResult<T> {
      * @param message 提示信息
      */
     public static <T> CommonResult<T> failed(String message) {
-        return new CommonResult<T>(ResultCode.FAILED.getCode(), message, null);
+        return new CommonResult<T>(Long.valueOf(ResultCode.FAILED.getCode()).intValue(), message, null);
     }
 
     /**
@@ -90,28 +90,28 @@ public class CommonResult<T> {
      * @param message 提示信息
      */
     public static <T> CommonResult<T> validateFailed(String message) {
-        return new CommonResult<T>(ResultCode.VALIDATE_FAILED.getCode(), message, null);
+        return new CommonResult<T>(Long.valueOf(ResultCode.VALIDATE_FAILED.getCode()).intValue(), message, null);
     }
 
     /**
      * 未登录返回结果
      */
     public static <T> CommonResult<T> unauthorized(T data) {
-        return new CommonResult<T>(ResultCode.UNAUTHORIZED.getCode(), ResultCode.UNAUTHORIZED.getMessage(), data);
+        return new CommonResult<T>(Long.valueOf(ResultCode.UNAUTHORIZED.getCode()).intValue(), ResultCode.UNAUTHORIZED.getMessage(), data);
     }
 
     /**
      * 未授权返回结果
      */
     public static <T> CommonResult<T> forbidden(T data) {
-        return new CommonResult<T>(ResultCode.FORBIDDEN.getCode(), ResultCode.FORBIDDEN.getMessage(), data);
+        return new CommonResult<T>(Long.valueOf(ResultCode.FORBIDDEN.getCode()).intValue(), ResultCode.FORBIDDEN.getMessage(), data);
     }
 
     public long getCode() {
         return code;
     }
 
-    public void setCode(long code) {
+    public void setCode(int code) {
         this.code = code;
     }
